Advance
So, currently, everything almost works.

I feel like this is an important version I needed to commit. However, it's not a finished version yet, so I can't think of a good name for it yet.

Thus, I'm just calling it "advance" as in advancing forward another step.

So the problem now, is that I need to understand the things on a deeper level.

Anyway, rest of the work is tomorrow. So, commit and sleep.